Upgrade react (#52)
* Upgrade dependencies

* Upgrade cra

* Upgrade to react 18

* Revert "Upgrade to react 18"

This reverts commit 3032dc04ee340c089d9ab9510531497a035947b6.

* Upgrade to mui (1)

* Upgrade to material-ui

* Fix build warning

One of your dependencies, babel-preset-react-app, is importing the
"@babel/plugin-proposal-private-property-in-object" package without
declaring it in its dependencies. This is currently working because
"@babel/plugin-proposal-private-property-in-object" is already in your
node_modules folder for unrelated reasons, but it may break at any time.

babel-preset-react-app is part of the create-react-app project, which
is not maintianed anymore. It is thus unlikely that this bug will
ever be fixed. Add "@babel/plugin-proposal-private-property-in-object" to
your devDependencies to work around this error. This will make this message
go away.

* Fix migration errors

* Fix react render & amplify import

* Fix styling

* Fix imports

* Fix deprecation warnings

* Debugging

* Working

Without play/pause/currentTime

* Split state into two

* Format sources

* Upgrade aws amplify

* Upgrade dependencies

* Upgrade infrastructure dependencies

* Update build script

* Trigger only on PR

* Downgrade typescript to fix build

* Suppress warning

* Split build steps

* Replace cache step

* Configure cache path

* Fix workflow

---------

Co-authored-by: kaklakariada <kaklakariada@users.noreply.github.com>